feat: add port duplex status ans speed unit as tags
Recent firmwares apparently display speed as something like '1000 Mbps
Full', where previously it was '1000 Mbps', causing the string parser to
not properly extract the speed.

We now split the port speed string on whitespace, making sure we have
the integer value, and use the other 1 or two parts as tags.